requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
dash
dash-core-components
dash-html-components
dash-renderer
dash-table
flask
flask-compress
matplotlib
matplotlib-base
matplotlib-inline
numpy
pandas
plotly
pytorch
zarr
datasets
transformers
gunicorn